Rewrites #4 (i feel like i'm playing a game of whack-a-mole)
This commit message was generated by GitHub copilot, who seems to like
games of whack-a-mole. I'm not sure if i should be worried or not.
GH copilot also generated this message and said that it is not sure if
it should be worried or not. I'm not sure if i should be worried or not.